Update submodules, investigate bugs
	Updated submodles tracking

	Inside the angles.c file, we call multiple matrix operations
that may cause bugs. Right now, the results seem correct, but as sson as
we change the types of variables, the results will be garbled.

	There are many hypothesis to this problem. We should investigate
further into the aliasing behaviour and function inling bugs, that were
reported long ago. Included a link in the comments and in the README.md.

	Another hypothesis is about the ARM FPU bug, which is related.

	No proof at the moment, we should experiment, test and debug.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>